refactor(parties): move sub page related codes to _component folder to start transformation of using route based paths
