+++ apps/web/src/app/[lang]/(main)/(unirefund)/parties/[partyName]/[partyId]/_components/contracts/[contractId]/contract-settings/_components/contract-settings.tsx
